requirements.txt
flask==2.0.1
flask-cors==3.0.10
textblob==0.15.3
python-docx==0.8.11
PyPDF2==3.0.1
nltk==3.6.3
reportlab==3.6.12
Werkzeug==2.0.3